requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==5.2.11
djangorestframework==3.16.1
pycparser==3.0
pyOpenSSL==25.3.0
sqlparse==0.5.5
typing_extensions==4.15.0
# uWSGI==2.0.31
gunicorn==25.0.3